### 2018-02-15 e7675b0
Minor improvement documentation kdv_fscatter
