### 2018-02-15 63f8445
Added SWSS splitting scheme
