+++ src/private/fnft__kdv_discretization.c
@@ -34,6 +34,7 @@ UINT fnft__kdv_discretization_degree(kdv_discretization_t
         case kdv_discretization_2SPLIT1B:
         case kdv_discretization_2SPLIT2A: // With change of base trick
         case kdv_discretization_2SPLIT2B:
+        case kdv_discretization_2SPLIT2S:
             return 1;
         case kdv_discretization_2SPLIT4B:
             return 2;
@@ -72,6 +73,7 @@ REAL fnft__kdv_discretization_boundary_coeff(kdv_discretization_t discretization
         case kdv_discretization_2SPLIT1B:
         case kdv_discretization_2SPLIT2A:
         case kdv_discretization_2SPLIT2B:
+        case kdv_discretization_2SPLIT2S:
         case kdv_discretization_2SPLIT3A:
         case kdv_discretization_2SPLIT3B:
         case kdv_discretization_2SPLIT4A:
