### 2018-02-15 7513ee6
Added third order symmetric scheme
