+++ src/private/fnft__kdv_fscatter.c
@@ -205,7 +205,7 @@ INT kdv_fscatter(const UINT D, COMPLEX const * const q,
             
             break;
             
-       case kdv_discretization_2SPLIT3A:
+        case kdv_discretization_2SPLIT3A:
 
             e_1B = &e_Bstorage[0];
             e_2B = &e_Bstorage[3];
@@ -280,8 +280,38 @@ INT kdv_fscatter(const UINT D, COMPLEX const * const q,
             }
 
             break;
-        
-      case kdv_discretization_2SPLIT4A:
+        case kdv_discretization_2SPLIT3S:
+            
+            e_1B = &e_Bstorage[0];
+            e_2B = &e_Bstorage[3];
+
+            for (i=D-1; i>=0; i--) {
+                
+                kdv_fscatter_zero_freq_scatter_matrix(e_1B, eps_t/ *deg_ptr, q[i]);
+                kdv_fscatter_zero_freq_scatter_matrix(e_2B, 2*eps_t/ *deg_ptr, q[i]);
+
+                // construct the scattering matrix for the i-th sample
+                p11[2] = 2*e_1B[1]*e_1B[2]/3;
+                p11[1] = 0.0;
+                p11[0] = (2*e_1B[0]*e_1B[0] + e_2B[0])/3;
+                p12[2] = (4*e_1B[0]*e_1B[1] - e_2B[1])/6;
+                p12[1] = 2*e_2B[1]/3;
+                p12[0] = p12[2];
+                p21[2] = (4*e_1B[0]*e_1B[2] - e_2B[2])/6;
+                p21[1] = 2*e_2B[2]/3;
+                p21[0] = p21[2];
+                p22[2] = p11[0];
+                p22[1] = 0.0;
+                p22[0] = p11[2];
+                
+                p11 += *deg_ptr + 1;
+                p21 += *deg_ptr + 1;
+                p12 += *deg_ptr + 1;
+                p22 += *deg_ptr + 1;
+            }
+            
+            break;
+        case kdv_discretization_2SPLIT4A:
 
             e_2B = &e_Bstorage[0];
             e_4B = &e_Bstorage[3];
