### 2018-02-22 f738429
Changing .travis.yml as described in https://docs.travis-ci.com/user/languages/c/. Will add more configs if this works.
