+++ .travis.yml
@@ -11,7 +11,7 @@ matrix:
             - gcc-5
             - gfortran-5
       env:
-        - MATRIX_EVAL="CC=gcc-5 && FC=gfortran-5"
+        - GCC_VER="5" MATRIX_EVAL="CC=gcc-5 && FC=gfortran-5"
 
 script: cmake . && make && make test
 
