### 2018-02-22 3bb002b
Giving up, reverting to old .travis.yml
