### 2018-02-24 d0843ad
akns_fscatter adapted to kdv. Tests for 5A and 5B fail.
