+++ src/private/fnft__akns_fscatter.c
@@ -1,0 +1,885 @@
+/*
+* This file is part of FNFT.  
+*                                                                  
+* FNFT is free software; you can redistribute it and/or
+* modify it under the terms of the version 2 of the GNU General
+* Public License as published by the Free Software Foundation.
+*
+* FNFT is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+* GNU General Public License for more details.
+*                                                                      
+* You should have received a copy of the GNU General Public License
+* along with this program. If not, see <http://www.gnu.org/licenses/>.
+*
+* Contributors:
+* Sander Wahls (TU Delft) 2017-2018.
+
+*/
+
+#define FNFT_ENABLE_SHORT_NAMES
+
+#include "fnft__errwarn.h"
+#include "fnft__poly_fmult.h"
+#include "fnft__akns_fscatter.h"
+#include "fnft__akns_discretization.h"
+#include "fnft__discretization_t.h"
+#include "fnft__misc.h"
+
+/**
+ * Returns the length of array to be allocated based on the number
+ * of samples and discretization.
+ */
+UINT akns_fscatter_numel(UINT D, discretization_t discretization)
+{
+    // 2x2 matrix of degree+1 elements for each sample
+    return 4*(akns_discretization_degree(discretization) + 1)*D;
+}
+
+/**
+ * Returns the scattering matrix for a single step at frequency zero.
+ */
+INT fnft__akns_fscatter_zero_freq_scatter_matrix(COMPLEX *M,
+                                                const REAL eps_t, const REAL q, discretization_opts_t *opts)
+{   
+    REAL Delta = eps_t * CSQRT(q);
+
+    switch(opts->evolution_equation){
+	case evolution_equation_kdv:
+
+    	 M[0] = CCOS(Delta);                // M(1,1) = M(2,2)
+   	 M[2] = -eps_t * misc_CSINC(Delta); // M(2,1)
+    	 M[1] = -q * M[2];                  // M(1,2)
+         return 0;
+    	break;
+
+       default:
+        return E_INVALID_ARGUMENT(opts->contspec_type);
+    }
+}
+/**
+ * Fast computation of polynomial approximation of the combined scattering
+ * matrix.
+ */
+INT akns_fscatter(const UINT D, COMPLEX const * const q,
+                 const REAL eps_t, COMPLEX * const result, UINT * const deg_ptr,
+                 discretization_opts_t *opts)
+{
+    
+    INT i, ret_code, W;
+    COMPLEX *p, *p11, *p12, *p21, *p22;
+    UINT n, len;
+    COMPLEX e_Bstorage[21];
+    COMPLEX *e_0_5B, *e_1B, *e_1_5B, *e_2B, *e_3B, *e_4B, *e_5B, *e_6B, *e_8B,
+                                        *e_10B, *e_12B, *e_15B, *e_21B, *e_24B,
+                                        *e_30B, *e_35B, *e_42B, *e_70B, *e_105B;
+
+    // Check inputs
+    if (D == 0)
+        return E_INVALID_ARGUMENT(D);
+    if (q == NULL)
+        return E_INVALID_ARGUMENT(u);
+    if (eps_t <= 0.0)
+        return E_INVALID_ARGUMENT(eps_t);
+    if (result == NULL)
[…]

[…]
