### 2018-02-24 #582 @wahls
@kyleniemeyer, Hartmut @hafermann would be a good choice if he is available. He also does research in the field and has experience with C++ open source development.
