### 2018-02-25 36e1990
Partial progress with adapting nse_fscatter to akns_fscatter.
