### 2018-02-26 cca5d2a
Bugfix: Wrong mapping in kdv_fscatter for 2split5A and 2split5B
All KdV test pass now.
