### 2018-02-27 #582 @hafermann
If not too urgent, I would be happy to review the software within, say, three weeks from now. I intended to try it in any case.
