### 2018-02-27 #597 @whedon
[REVIEW]: FNFT: A Software Library for Computing Nonlinear Fourier Transforms
**Submitting author:** @wahls (<a href="http://orcid.org/0000-0001-5159-0918">Sander Wahls</a>)
**Repository:** <a href="https://github.com/FastNFT/FNFT" target ="_blank">https://github.com/FastNFT/FNFT</a>
**Version:** v0.1.0
