### 2018-03-02 f89a063
Bugfix: KdV for potentials containing (a) negative value(s)
Issue: The intermediate value Delta in kdv_fscatter_zero_freq_scatter_matrix was declared as REAL, but should contain an imaginary value for negative potential samples.
Fix: Changed the declaration of Delta to COMPLEX and added a test case with a negative potential sample.
