### 2018-03-02 78b5c9b
Progress with adapting nse functions to akns functions. MODAL yet to be adapted.
