+++ src/private/fnft__akns_fscatter.c
@@ -40,60 +40,29 @@ UINT akns_fscatter_numel(UINT D, discretization_t discretization)
 /**
  * Returns the scattering matrix for a single step at frequency zero.
  */
-INT fnft__akns_fscatter_zero_freq_scatter_matrix(COMPLEX *M,
-                                                const REAL eps_t, const COMPLEX q, discretization_opts_t *opts)
+static inline void akns_fscatter_zero_freq_scatter_matrix(COMPLEX * const M,
+                                                const REAL eps_t, const COMPLEX q, const COMPLEX r)
 {   
-    REAL Delta;
-    COMPLEX r;
-
-    switch(opts->evolution_equation){
-	case evolution_equation_kdv:
-         
-         Delta = eps_t * CSQRT(q);
-    	 M[0] = CCOS(Delta);                // M(1,1) = M(2,2)
-   	 M[2] = -eps_t * misc_CSINC(Delta); // M(2,1)
-    	 M[1] = -q * M[2];                  // M(1,2)
-         return 0;
-    	break;
-	
-	case evolution_equation_fnse:
-
-         r = -CONJ(q);
-         if (q == 0){
-             M[0] = 1; M[1] = 0; M[2] = 0;
-         }else{
-             M[0] = CCOSH(eps_t*CSQRT(q)*CSQRT(r));
-             M[1] = (CSINH(eps_t*CSQRT(r*q))*CSQRT(r*q))/r;
-             M[2] = -CONJ(M[1]);
-         }
-
-	case evolution_equation_dnse:
-
-         r = CONJ(q);
-         if (q == 0){
-             M[0] = 1; M[1] = 0; M[2] = 0;
-         }else{
-             M[0] = CCOSH(eps_t*CSQRT(q)*CSQRT(r));
-             M[1] = (CSINH(eps_t*CSQRT(r*q))*CSQRT(r*q))/r;
-             M[2] = CONJ(M[1]);
-         }
-       default:
-        return E_INVALID_ARGUMENT(opts->contspec_type);
-    }
+    COMPLEX Delta, del;
+    Delta = eps_t * CSQRT(-q*r);
+    del = eps_t * misc_CSINC(Delta);
+    M[0] = CCOS(Delta);
+    M[2] = r * del;
+    M[1] = q * del;
 }
 /**
  * Fast computation of polynomial approximation of the combined scattering
  * matrix.
  */
-INT akns_fscatter(const UINT D, COMPLEX const * const q,
+INT akns_fscatter(const UINT D, COMPLEX const * const q, COMPLEX const * const r,
                  const REAL eps_t, COMPLEX * const result, UINT * const deg_ptr,
-                 INT * const W_ptr, discretization_opts_t *opts)
+                 INT * const W_ptr, discretization_t discretization)
 {
     
     INT i, ret_code;
     COMPLEX *p, *p11, *p12, *p21, *p22;
     UINT n, len;
-    COMPLEX e_Bstorage[21];
+    COMPLEX e_Bstorage[21], scl;
     COMPLEX *e_0_5B, *e_1B, *e_1_5B, *e_2B, *e_3B, *e_4B, *e_5B, *e_6B, *e_8B,
                                         *e_10B, *e_12B, *e_15B, *e_21B, *e_24B,
                                         *e_30B, *e_35B, *e_42B, *e_70B, *e_105B;
@@ -102,7 +71,9 @@ INT akns_fscatter(const UINT D, COMPLEX const * const q,
     if (D == 0)
         return E_INVALID_ARGUMENT(D);
     if (q == NULL)
-        return E_INVALID_ARGUMENT(u);
[…]
