### 2018-03-02 184d8df
Common function akns_scatter_matrix created. Adapted to nse and kdv. Test case added for kdv.
