+++ include/fnft_kdv_discretization_t.h
@@ -83,7 +83,8 @@ typedef enum {
     fnft_kdv_discretization_2SPLIT7A,
     fnft_kdv_discretization_2SPLIT7B,
     fnft_kdv_discretization_2SPLIT8A,
-    fnft_kdv_discretization_2SPLIT8B
+    fnft_kdv_discretization_2SPLIT8B,
+    fnft_kdv_discretization_BO
 } fnft_kdv_discretization_t;
 
 #ifdef FNFT_ENABLE_SHORT_NAMES
@@ -105,6 +106,7 @@ typedef enum {
 #define kdv_discretization_2SPLIT7B fnft_kdv_discretization_2SPLIT7B
 #define kdv_discretization_2SPLIT8A fnft_kdv_discretization_2SPLIT8A
 #define kdv_discretization_2SPLIT8B fnft_kdv_discretization_2SPLIT8B
+#define kdv_discretization_BO fnft_kdv_discretization_BO
 #define kdv_discretization_t fnft_kdv_discretization_t
 #endif
 
