+++ include/private/fnft__nse_scatter.h
@@ -28,6 +28,7 @@
 #define FNFT__NSE_SCATTER_H
 
 #include "fnft_nse_discretization_t.h"
+#include "fnft__akns_scatter.h"
 
 /**
  * @brief Computes \f$a(\lambda)\f$, \f$ a'(\lambda) = \frac{\partial a(\lambda)}{\partial \lambda}\f$
