### 2018-03-08 8167823
poly_fmult can now multiply arbitrary numbers of polynomials (note: if n is not a power of two, than p has to be larger than the number of inputs coefficients)
