+++ test/fnft__poly/fnft__poly_fmult_test.c
@@ -23,30 +23,7 @@
 #include "fnft__misc.h"
 #include "fnft__errwarn.h"
 
-/*INT test_poly_fmult2()
-{
-    const INT deg = 4;
-    kiss_fft_cpx p1[5] = {1, 0, 0, 2, 3, 0, 0, 4, 5, 0};
-    kiss_fft_cpx p2[5] = {0, 5, 6, 0, 0, 7, 8, 0, 0, 9};
-    kiss_fft_cpx result[9];
-    kiss_fft_cpx result_exact[9] = {0, 5, -4, 0, 0, 34, -8, 0, 0, 95, 8, 0, \
-        0, 94, 4, 0, 0, 45};
-    void *mem;
-    UINT lenmem;
-
-    lenmem = poly_fmult2_lenmen(deg);
-    mem = malloc(lenmem);
-    if (mem == NULL)
-        return E_NOMEM;
-
-    if (poly_fmult2(deg, p1, p2, result, mem, lenmem, 0) == FNFT_ERROR)
-        return E_SUBROUTINE;
-    if (rel_err(2*deg+1, result, result_exact) > 10*DBL_EPSILON)
-        return FNFT_E_INCORRECT_RESULT;
-    return SUCCESS;
-}*/
-
-static INT poly_fmult_test(INT normalize_flag)
+static INT poly_fmult_test_D_is_power_of_2(INT normalize_flag)
 {
     UINT deg = 2, n = 8, i;
     COMPLEX p[24];
@@ -56,23 +33,81 @@ static INT poly_fmult_test(INT normalize_flag)
         format long g; r.'
     */
     COMPLEX result_exact[18] = { \
-           10690.1093069785 - I*7285.45565246217, \
-           22605.6811426431 - I*19510.065631389, \
-           57389.5004107115 - I*20763.2575587549, \
-           53847.0938223071 - I*18963.5560762862, \
-           88648.7268642591 - I*2309.98190237675, \
-           66515.8698354361 - I*3878.11603829108, \
-           45160.5175520559 + I*12083.5337878339, \
-           64021.1527510823 + I*34487.976136233, \
-          -26614.7840829275 + I*18912.0673174305, \
-           624.669240812309 + I*18294.2454730283, \
-           7755.19752834515 + I*18447.5313133858, \
-          -33812.2918248361 + I*1257.46809193287, \
-          -711.103319555697 - I*2023.66741077336, \
-           9781.54461527385 + I*407.75394285625, \
-           39.3498537600159 + I*420.245708764338, \
-          -508.134987301475 + I*156.212629759735, \
+           10690.1093069785 - I*7285.45565246217,
+           22605.6811426431 - I*19510.065631389, 
+           57389.5004107115 - I*20763.2575587549,
+           53847.0938223071 - I*18963.5560762862,
+           88648.7268642591 - I*2309.98190237675,
+           66515.8698354361 - I*3878.11603829108,
+           45160.5175520559 + I*12083.5337878339,
+           64021.1527510823 + I*34487.976136233, 
+          -26614.7840829275 + I*18912.0673174305,
+           624.669240812309 + I*18294.2454730283,
+           7755.19752834515 + I*18447.5313133858,
+          -33812.2918248361 + I*1257.46809193287,
+          -711.103319555697 - I*2023.66741077336,
+           9781.54461527385 + I*407.75394285625, 
+           39.3498537600159 + I*420.245708764338,
+          -508.134987301475 + I*156.212629759735,
            4.42978955469815 - I*0.0813363246183392 };
+    INT W, *W_ptr = NULL;
+    REAL scl;
+    INT ret_code;
+
+    for (i=0; i<(deg+1)*n; i++) {
[…]
