### 2018-03-10 b719bf1
Added function misc_nextpowerof2(...)
