### 2018-03-10 c7a89ba
poly_fmult2x2 can now multiply arbitrary numbers of polynomials
