+++ test/fnft__poly/fnft__poly_fmult_test.c
@@ -23,16 +23,16 @@
 #include "fnft__misc.h"
 #include "fnft__errwarn.h"
 
-static INT poly_fmult_test_D_is_power_of_2(INT normalize_flag)
+static INT poly_fmult_test_n_is_power_of_2(INT normalize_flag)
 {
     UINT deg = 2, n = 8, i;
-    COMPLEX p[24];
+    COMPLEX p[n*(deg + 1)]; // since n is already a power of two
     /* Matlab code to generate the "exact" result:
         i=0:23; p=sqrt(i+1).*(cos(i) + 1j*sin(-2*i));
         r=p(1:3); for (k=4:3:24), r=conv(r,p(k:(k+2))); end
         format long g; r.'
     */
-    COMPLEX result_exact[18] = { \
+    COMPLEX result_exact[17] = {
            10690.1093069785 - I*7285.45565246217,
            22605.6811426431 - I*19510.065631389, 
            57389.5004107115 - I*20763.2575587549,
@@ -50,6 +50,7 @@ static INT poly_fmult_test_D_is_power_of_2(INT normalize_flag)
            39.3498537600159 + I*420.245708764338,
           -508.134987301475 + I*156.212629759735,
            4.42978955469815 - I*0.0813363246183392 };
+    const UINT deg_exact = sizeof(result_exact)/sizeof(result_exact[0]) - 1;
     INT W, *W_ptr = NULL;
     REAL scl;
     INT ret_code;
@@ -62,6 +63,8 @@ static INT poly_fmult_test_D_is_power_of_2(INT normalize_flag)
     ret_code = poly_fmult(&deg, n, p, W_ptr);
     if (ret_code != SUCCESS)
         return E_SUBROUTINE(ret_code);
+    if (deg != deg_exact)
+        return E_TEST_FAILED;
     if (normalize_flag) {
         if (W == 0)
             return E_TEST_FAILED;
@@ -74,10 +77,10 @@ static INT poly_fmult_test_D_is_power_of_2(INT normalize_flag)
     return SUCCESS;
 }
 
-static INT poly_fmult_test_D_is_no_power_of_2(INT normalize_flag)
+static INT poly_fmult_test_n_is_no_power_of_2(INT normalize_flag)
 {
     UINT deg = 2, n = 11, i;
-    COMPLEX p[(deg+1)*16]; // 16 = next power of two of n=11
+    COMPLEX p[(deg+1)*misc_nextpowerof2(n)];
     /* Matlab code to generate the "exact" result:
         deg=2; n=11; N=(deg+1)*n; i=0:(N-1);
         p=sqrt(i+1).*(cos(i) + 1j*sin(-2*i)); r=p(1:(deg+1));
@@ -108,6 +111,7 @@ static INT poly_fmult_test_D_is_no_power_of_2(INT normalize_flag)
             50207.974399346 -      84587.5774434785*I,
            56394.1550202597 -      147664.217837454*I,
           -801.466849402119 +      1048.08952721767*I };        
+    const UINT deg_exact = sizeof(result_exact)/sizeof(result_exact[0]) - 1;
     INT W, *W_ptr = NULL;
     REAL scl;
     INT ret_code;
@@ -120,6 +124,8 @@ static INT poly_fmult_test_D_is_no_power_of_2(INT normalize_flag)
     ret_code = poly_fmult(&deg, n, p, W_ptr);
     if (ret_code != SUCCESS)
         return E_SUBROUTINE(ret_code);
+    if (deg != deg_exact)
+        return E_TEST_FAILED;
     if (normalize_flag) {
         if (W == 0)
[…]
