+++ src/private/fnft__poly_fmult.c
@@ -28,18 +28,18 @@
 #include "kiss_fft.h"
 #include "_kiss_fft_guts.h"
 
-static INT poly_fmult2_len(UINT deg)
+static inline INT poly_fmult2_len(UINT deg)
 {
     return kiss_fft_next_fast_size(2*(deg + 1) - 1);
 }
 
-static UINT poly_fmult2_lenmen(UINT deg)
+static inline UINT poly_fmult2_lenmen(UINT deg)
 {
     return sizeof(kiss_fft_cpx)*(4*poly_fmult2_len(deg) - 1);
 }
 
-static INT poly_fmult2(const UINT deg, COMPLEX *p1, \
-    COMPLEX *p2, COMPLEX *result, void *mem, \
+static inline INT poly_fmult2(const UINT deg, COMPLEX *p1, 
+    COMPLEX *p2, COMPLEX *result, void *mem, 
     kiss_fft_cfg cfg_fft, kiss_fft_cfg cfg_ifft, INT add_flag)
 {
     UINT i, len;
@@ -89,7 +89,7 @@ static INT poly_fmult2(const UINT deg, COMPLEX *p1, \
     return SUCCESS;
 }
 
-static INT poly_rescale(const UINT d, COMPLEX * const p)
+static inline INT poly_rescale(const UINT d, COMPLEX * const p)
 {
     UINT i;
     INT a;
@@ -133,7 +133,8 @@ INT fnft__poly_fmult(UINT * const d, UINT n, COMPLEX * const p,
     lenmem = poly_fmult2_lenmen(deg * n);
     mem = malloc(lenmem); // contains the memory for the actual data
     // The line below find max number of bytes needed for an (I)FFT config
-    kiss_fft_alloc(poly_fmult2_lenmen(deg * n/2), 0, NULL, &memneeded);
+    kiss_fft_alloc(poly_fmult2_len(deg * misc_nextpowerof2(n)/2), 0, NULL,
+        &memneeded);
     mem_fft = malloc(memneeded); // memory for FFT configs
     mem_ifft = malloc(memneeded); // memory for IFFT configs
     if (mem == NULL || mem_fft == NULL || mem_ifft == NULL) {
@@ -180,7 +181,7 @@ INT fnft__poly_fmult(UINT * const d, UINT n, COMPLEX * const p,
         // polynomial -> simply keep it, but extend to next degree 2*deg+1
         if (mod_n != 0) {
 
-            INT diff = (n - mod_n)/2;
+            const INT diff = (n - mod_n)/2; // = p1 - result
             memmove(p1 + (deg - diff), p1, (deg+1)*sizeof(COMPLEX));
             for (i=0; i<deg; i++)
                 result[i] = 0;
@@ -210,7 +211,7 @@ release_mem:
     return ret_code;
 }
 
-static INT poly_rescale2x2(const UINT d,
+static inline INT poly_rescale2x2(const UINT d,
     COMPLEX * const p11,
     COMPLEX * const p12,
     COMPLEX * const p21,
@@ -256,15 +257,10 @@ static INT poly_rescale2x2(const UINT d,
     return a;
 }
 
-/*
-* length of p = m*m*n*(deg+1)
-* length of result = m*m*(n/2)*(2*deg+1)
-* WARNING: p is overwritten
-*/
 INT fnft__poly_fmult2x2(UINT * const d, UINT n, COMPLEX * const p,
     COMPLEX * const result, INT * const W_ptr)
 {
-    UINT i, deg, lenmem, len, memneeded, memneeded_buf;
+    UINT i, deg, lenmem, len, memneeded, memneeded_buf, deg_excess = 0;
[…]
