### 2018-03-11 a653cc0
Added functions to inform user about how much memory needs to be reserved for inputs p (and result) to poly_fmult(2x2)
