### 2018-03-23 #597 @wahls
Dear @hafermann, many thanks for reviewing. I have added the missing DOI's to the paper.
