### 2018-03-26 #597 @hafermann
Hi @kyleniemeyer, I was not sure whether comments were solicited looking at the reviewer guidelines. I have no further comments other than that the code submitted by @wahls compiled fine including all features, examples and documentation and worked out of the box. We have verified it also for independent test cases. It will for sure be very useful for the community.
