### 2018-03-30 #597 @kyleniemeyer
Hi @wahls, just a few changes for the paper:

- The reference to Turitsyn et al. 2017 in the second paragraph should be changed from "(Turitsyn et al. 2017)" to "Turitsyn et al. (2017)" by changing `[@Turitsyn2017]` to `@Turitsyn2017`
- In the second paragraph, change `[@Osborne2010], [@Bruehl2016]` to `[@Osborne2010;@Bruehl2016]` (this will put them in one set of parentheses)
- In the third paragraph, you should remove the "in" prior to the Wahls and Poor 2013 reference, and change the references to `[@Wahls2013;@Wahls2015;@Prins2018]`

Once you have made these changes, please archive your entire repository (e.g., using Zenodo) and give the DOI here.
